LoginActivity now fragmented
